Kernel loading
Added a basic kernel loading system that uses the FileFox32 FS. Expects a Kernel.bin entry within the first Root Table Sector.

There will be a later requirement to change some of the calculations performed to ensure the system works in all use cases, however for now it works to load the necessary information using the values provided within the FS Root.