Protected mode
Started implementing support for switching to protected mode. The existing RedFox32 kernel will be used as the sample for the bootloader.

There is now a dependancy on FileFox32 producing tools to interact with the RedFox32 OS, allowing for the writing of files, to simplify the writing of the kernel and other desired files to the disk after compilation, rather than building in fixed values.